feat: 전역 상태 객체에 장바구니 관련 값 추가 및 ui 출력 함수를 render 폴더로 분리
